@@ -106,7 +106,7 @@ pub fn parse_file(src: &str, file_id: FileID) -> Result<AST, Box<Report>> {
                 message: format!("Extra token `{}` found.", token.2),
                 location: token.0..token.2,
             },
-            _ => ParsingError { file_id, message: format!("{}", parse_error), location: 0..0 },
+            _ => ParsingError { file_id, message: format!("{parse_error}"), location: 0..0 },
         })
         .map_err(|error| Box::new(error.into_report()))
 }
@@ -114,6 +114,16 @@ pub fn parse_file(src: &str, file_id: FileID) -> Result<AST, Box<Report>> {
 pub fn parse_string(src: &str) -> Option<AST> {
     let src = preprocess(src, 0).ok()?;
     lang::ParseAstParser::new().parse(&src).ok()
+}
+
+/// Parse a single (function or template) definition for testing purposes.
+use program_structure::ast::Definition;
+
+pub fn parse_definition(src: &str) -> Option<Definition> {
+    match parse_string(src) {
+        Some(AST { mut definitions, .. }) if definitions.len() == 1 => definitions.pop(),
+        _ => None,
+    }
 }
 
 #[must_use]
@@ -135,7 +145,7 @@ fn format_expected(tokens: &[String]) -> String {
             })
             .collect::<Vec<_>>()
             .join("");
-        format!(" Expected one of {}.", tokens)
+        format!(" Expected one of {tokens}.")
     }
 }
 
