@@ -1,13 +1,13 @@
-mod config;
-mod analysis_runner;
-
+use std::collections::HashSet;
 use std::path::PathBuf;
 use std::process::ExitCode;
 use clap::{CommandFactory, Parser};
 
-use analysis_runner::AnalysisRunner;
+use program_analysis::config;
+use program_analysis::analysis_runner::AnalysisRunner;
 
 use program_structure::constants::Curve;
+use program_structure::file_definition::FileID;
 use program_structure::report::Report;
 use program_structure::report::MessageCategory;
 use program_structure::writers::{LogWriter, ReportWriter, SarifWriter, CachedStdoutWriter};
@@ -35,9 +35,15 @@ struct Cli {
     #[clap(short = 'v', long = "verbose")]
     verbose: bool,
 
-    /// Set curve (BN128, BLS12_381, or GOLDILOCKS)
+    /// Set curve (BN254, BLS12_381, or GOLDILOCKS)
     #[clap(short = 'c', long = "curve", name = "NAME", default_value = config::DEFAULT_CURVE)]
     curve: Curve,
+}
+
+/// Returns true if a primary location of the report corresponds to a file
+/// specified on the command line by the user.
+fn filter_by_file(report: &Report, user_inputs: &HashSet<FileID>) -> bool {
+    report.primary_file_ids().iter().any(|file_id| user_inputs.contains(file_id))
 }
 
 /// Returns true if the report level is greater than or equal to the given
@@ -62,23 +68,30 @@ fn main() -> ExitCode {
         }
     }
 
-    // Set up analysis runner and analyze all functions and templates.
-    let allow_list = options.allow_list.clone();
-    let mut stdout_writer = CachedStdoutWriter::new(options.verbose)
-        .add_filter(move |report: &Report| filter_by_id(report, &allow_list))
-        .add_filter(move |report: &Report| filter_by_level(report, &options.output_level));
-    let mut runner = AnalysisRunner::new(&options.curve);
-    runner.with_files(&options.input_files, &mut stdout_writer);
+    // Set up analysis runner.
+    let (mut runner, reports) = AnalysisRunner::new(options.curve).with_files(&options.input_files);
 
-    runner.analyze_functions(&mut stdout_writer);
-    runner.analyze_templates(&mut stdout_writer);
+    // Set up writer and write reports to `stdout`.
+    let allow_list = options.allow_list.clone();
+    let user_inputs = runner.file_library().user_inputs().clone();
+    let mut stdout_writer = CachedStdoutWriter::new(options.verbose)
+        .add_filter(move |report: &Report| filter_by_level(report, &options.output_level))
+        .add_filter(move |report: &Report| filter_by_file(report, &user_inputs))
+        .add_filter(move |report: &Report| filter_by_id(report, &allow_list));
+    stdout_writer.write_reports(&reports, runner.file_library());
+
+    // Analyze functions and templates in user provided input files.
+    runner.analyze_functions(&mut stdout_writer, true);
+    runner.analyze_templates(&mut stdout_writer, true);
 
     // If a Sarif file is passed to the program we write the reports to it.
     if let Some(sarif_file) = options.sarif_file {
         let allow_list = options.allow_list.clone();
+        let user_inputs = runner.file_library().user_inputs().clone();
         let mut sarif_writer = SarifWriter::new(&sarif_file)
-            .add_filter(move |report: &Report| filter_by_id(report, &allow_list))
-            .add_filter(move |report: &Report| filter_by_level(report, &options.output_level));
+            .add_filter(move |report: &Report| filter_by_level(report, &options.output_level))
+            .add_filter(move |report: &Report| filter_by_file(report, &user_inputs))
+            .add_filter(move |report: &Report| filter_by_id(report, &allow_list));
         if sarif_writer.write_reports(stdout_writer.reports(), runner.file_library()) > 0 {
             stdout_writer.write_message(&format!("Result written to `{}`.", sarif_file.display()));
         }
