@@ -1,12 +1,9 @@
-use log::debug;
+use log::{debug, trace};
 use std::path::PathBuf;
 use std::collections::HashMap;
 
 use parser::ParseResult;
-use program_analysis::{
-    analysis_context::{AnalysisContext, AnalysisError},
-    get_analysis_passes,
-};
+
 use program_structure::{
     writers::{LogWriter, ReportWriter},
     template_data::TemplateInfo,
@@ -17,55 +14,77 @@ use program_structure::{
     report::{ReportCollection, Report},
 };
 
-use crate::config;
+#[cfg(test)]
+use program_structure::template_library::TemplateLibrary;
+
+use crate::{
+    analysis_context::{AnalysisContext, AnalysisError},
+    get_analysis_passes, config,
+};
 
 type CfgCache = HashMap<String, Cfg>;
 type ReportCache = HashMap<String, ReportCollection>;
 
+/// A type responsible for caching CFGs and running analysis passes over all
+/// functions and templates.
+#[derive(Default)]
 pub struct AnalysisRunner {
     curve: Curve,
+    /// The corresponding file library including file includes.
     file_library: FileLibrary,
+    /// Template ASTs generated by the parser.
     template_asts: TemplateInfo,
+    /// Function ASTs generated by the parser.
     function_asts: FunctionInfo,
+    /// Cached template CFGs generated on demand.
     template_cfgs: CfgCache,
+    /// Cached function CFGs generated on demand.
     function_cfgs: CfgCache,
+    /// Reports created during CFG generation.
     template_reports: ReportCache,
+    /// Reports created during CFG generation.
     function_reports: ReportCache,
 }
 
 impl AnalysisRunner {
-    pub fn new(curve: &Curve) -> Self {
-        AnalysisRunner {
-            curve: curve.clone(),
-            file_library: FileLibrary::new(),
-            template_asts: TemplateInfo::new(),
-            function_asts: FunctionInfo::new(),
-            template_cfgs: CfgCache::new(),
-            function_cfgs: CfgCache::new(),
-            template_reports: ReportCache::new(),
-            function_reports: ReportCache::new(),
-        }
+    pub fn new(curve: Curve) -> Self {
+        AnalysisRunner { curve, ..Default::default() }
     }
 
-    pub fn with_files(
-        &mut self,
-        input_files: &[PathBuf],
-        writer: &mut (impl LogWriter + ReportWriter),
-    ) -> &mut Self {
-        let (template_asts, function_asts, file_library) =
-            match parser::parse_files(input_files, &config::COMPILER_VERSION) {
-                ParseResult::Program(program, warnings) => {
-                    writer.write_reports(&warnings, &program.file_library);
-                    (program.templates, program.functions, program.file_library)
-                }
-                ParseResult::Library(library, warnings) => {
-                    writer.write_reports(&warnings, &library.file_library);
-                    (library.templates, library.functions, library.file_library)
-                }
-            };
-        self.template_asts = template_asts;
-        self.function_asts = function_asts;
-        self.file_library = file_library;
+    pub fn with_files(mut self, input_files: &[PathBuf]) -> (Self, ReportCollection) {
+        let reports = match parser::parse_files(input_files, &config::COMPILER_VERSION) {
+            ParseResult::Program(program, warnings) => {
+                self.template_asts = program.templates;
+                self.function_asts = program.functions;
+                self.file_library = program.file_library;
+                warnings
+            }
+            ParseResult::Library(library, warnings) => {
+                self.template_asts = library.templates;
+                self.function_asts = library.functions;
+                self.file_library = library.file_library;
+                warnings
+            }
+        };
+        (self, reports)
+    }
+
+    /// Convenience method used to generate a runner for testing purposes.
+    #[cfg(test)]
+    pub fn with_src(mut self, file_contents: &[&str]) -> Self {
+        use parser::parse_definition;
+
+        let mut library_contents = HashMap::new();
+        let mut file_library = FileLibrary::default();
+        for (file_index, file_source) in file_contents.iter().enumerate() {
+            let file_name = format!("file-{file_index}.circom");
+            let file_id = file_library.add_file(file_name, file_source.to_string(), true);
+            library_contents.insert(file_id, vec![parse_definition(file_source).unwrap()]);
+        }
+        let template_library = TemplateLibrary::new(library_contents, file_library.clone());
+        self.template_asts = template_library.templates;
+        self.function_asts = template_library.functions;
+        self.file_library = template_library.file_library;
 
         self
     }
@@ -74,14 +93,34 @@ impl AnalysisRunner {
         &self.file_library
     }
 
-    pub fn template_names(&self) -> Vec<String> {
+    pub fn template_names(&self, user_input_only: bool) -> Vec<String> {
         // Clone template names to avoid holding multiple references to `self`.
-        self.template_asts.keys().cloned().collect()
+        self.template_asts
+            .iter()
+            .filter_map(|(name, ast)| {
+                if !user_input_only || self.file_library.is_user_input(ast.get_file_id()) {
+                    Some(name)
+                } else {
+                    None
+                }
+            })
+            .cloned()
+            .collect()
     }
 
-    pub fn function_names(&self) -> Vec<String> {
+    pub fn function_names(&self, user_input_only: bool) -> Vec<String> {
         // Clone function names to avoid holding multiple references to `self`.
-        self.function_asts.keys().cloned().collect()
+        self.function_asts
+            .iter()
+            .filter_map(|(name, ast)| {
+                if !user_input_only || self.file_library.is_user_input(ast.get_file_id()) {
+                    Some(name)
+                } else {
+                    None
+                }
+            })
+            .cloned()
+            .collect()
     }
 
     fn analyze_template<W: LogWriter + ReportWriter>(&mut self, name: &str, writer: &mut W) {
@@ -90,7 +129,7 @@ impl AnalysisRunner {
         // We take ownership of the CFG and any previously generated reports
         // here to avoid holding multiple mutable and immutable references to
         // `self`. This may lead to the CFG being regenerated during analysis if
-        // the template is invoked recursively.
+        // the template is invoked recursively. If it is then ¯\_(ツ)_/¯.
         let mut reports = self.take_template_reports(name);
         if let Ok(cfg) = self.take_template(name) {
             for analysis_pass in get_analysis_passes() {
@@ -104,8 +143,12 @@ impl AnalysisRunner {
         writer.write_reports(&reports, &self.file_library);
     }
 
-    pub fn analyze_templates<W: LogWriter + ReportWriter>(&mut self, writer: &mut W) {
-        for name in self.template_names() {
+    pub fn analyze_templates<W: LogWriter + ReportWriter>(
+        &mut self,
+        writer: &mut W,
+        user_input_only: bool,
+    ) {
+        for name in self.template_names(user_input_only) {
             self.analyze_template(&name, writer);
         }
     }
@@ -116,7 +159,7 @@ impl AnalysisRunner {
         // We take ownership of the CFG and any previously generated reports
         // here to avoid holding multiple mutable and immutable references to
         // `self`. This may lead to the CFG being regenerated during analysis if
-        // the function is invoked recursively.
+        // the function is invoked recursively. If it is then ¯\_(ツ)_/¯.
         let mut reports = self.take_function_reports(name);
         if let Ok(cfg) = self.take_function(name) {
             for analysis_pass in get_analysis_passes() {
@@ -130,8 +173,12 @@ impl AnalysisRunner {
         writer.write_reports(&reports, &self.file_library);
     }
 
-    pub fn analyze_functions<W: LogWriter + ReportWriter>(&mut self, writer: &mut W) {
-        for name in self.function_names() {
+    pub fn analyze_functions<W: LogWriter + ReportWriter>(
+        &mut self,
+        writer: &mut W,
+        user_input_only: bool,
+    ) {
+        for name in self.function_names(user_input_only) {
             self.analyze_function(&name, writer);
         }
     }
@@ -169,16 +216,19 @@ impl AnalysisRunner {
             }
             // Get the AST corresponding to the template.
             let Some(ast) = self.template_asts.get(name) else {
+                trace!("failed to lift unknown template `{name}`");
                 return Err(AnalysisError::UnknownTemplate { name: name.to_string() })
             };
             // Generate the template CFG from the AST. Cache any reports.
             let mut reports = ReportCollection::new();
             let cfg = generate_cfg(ast, &self.curve, &mut reports).map_err(|report| {
                 reports.push(*report);
+                trace!("failed to lift template `{name}`");
                 AnalysisError::FailedToLiftTemplate { name: name.to_string() }
             })?;
             self.append_template_reports(name, &mut reports);
             self.template_cfgs.insert(name.to_string(), cfg);
+            trace!("successfully lifted template `{name}`");
         }
         Ok(self.template_cfgs.get(name).unwrap())
     }
@@ -192,44 +242,45 @@ impl AnalysisRunner {
             }
             // Get the AST corresponding to the function.
             let Some(ast) = self.function_asts.get(name) else {
+                trace!("failed to lift unknown function `{name}`");
                 return Err(AnalysisError::UnknownFunction { name: name.to_string() })
             };
             // Generate the function CFG from the AST. Cache any reports.
             let mut reports = ReportCollection::new();
             let cfg = generate_cfg(ast, &self.curve, &mut reports).map_err(|report| {
                 reports.push(*report);
+                trace!("failed to lift function `{name}`");
                 AnalysisError::FailedToLiftFunction { name: name.to_string() }
             })?;
             self.append_function_reports(name, &mut reports);
             self.function_cfgs.insert(name.to_string(), cfg);
+            trace!("successfully lifted function `{name}`");
         }
         Ok(self.function_cfgs.get(name).unwrap())
     }
 
-    fn take_template(&mut self, name: &str) -> Result<Cfg, AnalysisError> {
+    pub fn take_template(&mut self, name: &str) -> Result<Cfg, AnalysisError> {
         self.cache_template(name)?;
         // The CFG must be available since caching was successful.
         Ok(self.template_cfgs.remove(name).unwrap())
     }
 
-    fn take_function(&mut self, name: &str) -> Result<Cfg, AnalysisError> {
+    pub fn take_function(&mut self, name: &str) -> Result<Cfg, AnalysisError> {
         self.cache_function(name)?;
         // The CFG must be available since caching was successful.
         Ok(self.function_cfgs.remove(name).unwrap())
     }
 
-    fn replace_template(&mut self, name: &str, cfg: Cfg) -> bool {
+    pub fn replace_template(&mut self, name: &str, cfg: Cfg) -> bool {
         self.template_cfgs.insert(name.to_string(), cfg).is_some()
     }
 
-    fn replace_function(&mut self, name: &str, cfg: Cfg) -> bool {
+    pub fn replace_function(&mut self, name: &str, cfg: Cfg) -> bool {
         self.function_cfgs.insert(name.to_string(), cfg).is_some()
     }
 }
 
 impl AnalysisContext for AnalysisRunner {
-    type Error = AnalysisError;
-
     fn is_template(&self, name: &str) -> bool {
         self.template_asts.get(name).is_some()
     }
@@ -238,11 +289,11 @@ impl AnalysisContext for AnalysisRunner {
         self.function_asts.get(name).is_some()
     }
 
-    fn template(&mut self, name: &str) -> Result<&Cfg, Self::Error> {
+    fn template(&mut self, name: &str) -> Result<&Cfg, AnalysisError> {
         self.cache_template(name)
     }
 
-    fn function(&mut self, name: &str) -> Result<&Cfg, Self::Error> {
+    fn function(&mut self, name: &str) -> Result<&Cfg, AnalysisError> {
         self.cache_function(name)
     }
 
@@ -250,7 +301,7 @@ impl AnalysisContext for AnalysisRunner {
         &self,
         file_id: &FileID,
         file_location: &FileLocation,
-    ) -> Result<String, Self::Error> {
+    ) -> Result<String, AnalysisError> {
         let Ok(file) = self.file_library.to_storage().get(*file_id) else {
             return Err(AnalysisError::UnknownFile { file_id: *file_id });
         };
@@ -278,14 +329,13 @@ fn generate_cfg<Ast: IntoCfg>(
 
 #[cfg(test)]
 mod tests {
-    use parser::parse_definition;
-    use program_structure::{template_library::TemplateLibrary, intermediate_representation::Statement};
+    use program_structure::ir::Statement;
 
     use super::*;
 
     #[test]
     fn test_function() {
-        let mut runner = runner_from_src(&[r#"
+        let mut runner = AnalysisRunner::new(Curve::Goldilocks).with_src(&[r#"
             function foo(a) {
                 return a[0] + a[1];
             }
@@ -314,7 +364,7 @@ mod tests {
 
     #[test]
     fn test_template() {
-        let mut runner = runner_from_src(&[r#"
+        let mut runner = AnalysisRunner::new(Curve::Goldilocks).with_src(&[r#"
             template Foo(n) {
                 signal input a[2];
 
@@ -346,7 +396,7 @@ mod tests {
     #[test]
     fn test_underlying_str() {
         use Statement::*;
-        let mut runner = runner_from_src(&[r#"
+        let mut runner = AnalysisRunner::new(Curve::Goldilocks).with_src(&[r#"
             template Foo(n) {
                 signal input a[2];
 
@@ -360,28 +410,11 @@ mod tests {
             let file_location = stmt.meta().file_location();
             let string = runner.underlying_str(&file_id, &file_location).unwrap();
             match stmt {
+                // TODO: Why do some statements include the semi-colon and others don't?
                 Declaration { .. } => assert_eq!(string, "signal input a[2]"),
                 ConstraintEquality { .. } => assert_eq!(string, "a[0] === a[1];"),
                 _ => unreachable!(),
             }
         }
-    }
-
-    fn runner_from_src(src: &[&str]) -> AnalysisRunner {
-        let mut file_content = HashMap::new();
-        let mut file_library = FileLibrary::default();
-        for (file_index, file_source) in src.iter().enumerate() {
-            let file_name = format!("{file_index}.circom");
-            let file_id = file_library.add_file(file_name, file_source.to_string());
-            println!("{file_id}");
-            file_content.insert(file_id, vec![parse_definition(file_source).unwrap()]);
-        }
-        let template_library = TemplateLibrary::new(file_content, file_library.clone());
-
-        let mut runner = AnalysisRunner::new(&Curve::Goldilocks);
-        runner.template_asts = template_library.templates;
-        runner.function_asts = template_library.functions;
-        runner.file_library = file_library;
-        runner
     }
 }
