@@ -11,6 +11,10 @@ pub struct Symbol {
     pub init: Option<Expression>,
 }
 
+pub struct TupleInit {
+    pub tuple_init: (AssignOp, Expression),
+}
+
 pub fn assign_with_op_shortcut(
     op: ExpressionInfixOpcode,
     meta: Meta,
@@ -56,7 +60,7 @@ pub fn split_declaration_into_single_nodes(
 
     for symbol in symbols {
         let with_meta = meta.clone();
-        let has_type = xtype;
+        let has_type = xtype.clone();
         let name = symbol.name.clone();
         let dimensions = symbol.is_array;
         let possible_init = symbol.init;
@@ -67,9 +71,9 @@ pub fn split_declaration_into_single_nodes(
             let substitution = build_substitution(meta.clone(), symbol.name, vec![], op, init);
             initializations.push(substitution);
         }
-        // If the variable is not initialialized it is default initialized to 0.
-        // We remove this because we don't want this assignment to be flagged as
-        // an unused assignment by the side-effect analysis.
+        // If the variable is not initialized it is default initialized to 0 by
+        // Circom.  We remove this because we don't want this assignment to be
+        // flagged as an unused assignment by the side-effect analysis.
         // else if xtype == Var {
         //     let mut value = Expression::Number(meta.clone(), BigInt::from(0));
         //     for dim_expr in dimensions.iter().rev() {
@@ -79,6 +83,51 @@ pub fn split_declaration_into_single_nodes(
         //     let substitution = build_substitution(meta.clone(), symbol.name, vec![], op, value);
         //     initializations.push(substitution);
         // }
+    }
+    build_initialization_block(meta, xtype, initializations)
+}
+
+pub fn split_declaration_into_single_nodes_and_multi_substitution(
+    meta: Meta,
+    xtype: VariableType,
+    symbols: Vec<Symbol>,
+    init: Option<TupleInit>,
+) -> Statement {
+    let mut initializations = Vec::new();
+    let mut values = Vec::new();
+    for symbol in symbols {
+        let with_meta = meta.clone();
+        let has_type = xtype.clone();
+        let name = symbol.name.clone();
+        let dimensions = symbol.is_array;
+        debug_assert!(symbol.init.is_none());
+        let single_declaration =
+            build_declaration(with_meta.clone(), has_type, name.clone(), dimensions.clone());
+        initializations.push(single_declaration);
+        // Circom default initializes local arrays to 0. We remove this because
+        // we don't want these assignments to be flagged as unused assignments
+        // by the side-effect analysis.
+        // if xtype == Var && init.is_none() {
+        //     let mut value = Expression::Number(meta.clone(), BigInt::from(0));
+        //     for dim_expr in dimensions.iter().rev() {
+        //         value = build_uniform_array(meta.clone(), value, dim_expr.clone());
+        //     }
+
+        //     let substitution =
+        //         build_substitution(meta.clone(), symbol.name, vec![], AssignOp::AssignVar, value);
+        //     initializations.push(substitution);
+        // }
+        values.push(Expression::Variable { meta: with_meta.clone(), name, access: Vec::new() })
+    }
+    if let Some(tuple) = init {
+        let (op, expression) = tuple.tuple_init;
+        let multi_sub = build_multi_substitution(
+            meta.clone(),
+            build_tuple(meta.clone(), values),
+            op,
+            expression,
+        );
+        initializations.push(multi_sub);
     }
     build_initialization_block(meta, xtype, initializations)
 }
