@@ -38,13 +38,40 @@ pub fn build_variable(meta: Meta, name: String, access: Vec<Access>) -> Expressi
 }
 
 pub fn build_number(meta: Meta, value: BigInt) -> Expression {
-    Expression::Number(meta, value)
+    Number(meta, value)
 }
 
 pub fn build_call(meta: Meta, id: String, args: Vec<Expression>) -> Expression {
     Call { meta, id, args }
 }
 
+pub fn build_anonymous_component(
+    meta: Meta,
+    id: String,
+    params: Vec<Expression>,
+    signals: Vec<Expression>,
+    names: Option<Vec<(AssignOp, String)>>,
+    is_parallel: bool,
+) -> Expression {
+    AnonymousComponent { meta, id, params, signals, names, is_parallel }
+}
+
 pub fn build_array_in_line(meta: Meta, values: Vec<Expression>) -> Expression {
     ArrayInLine { meta, values }
+}
+
+pub fn build_tuple(meta: Meta, values: Vec<Expression>) -> Expression {
+    Tuple { meta, values }
+}
+
+pub fn unzip_3(
+    vec: Vec<(String, AssignOp, Expression)>,
+) -> (Vec<(AssignOp, String)>, Vec<Expression>) {
+    let mut op_name = Vec::new();
+    let mut exprs = Vec::new();
+    for i in vec {
+        op_name.push((i.1, i.0));
+        exprs.push(i.2);
+    }
+    (op_name, exprs)
 }
